Try to make each file simpler
Goal was to follow the practice that a file should change for only one
reason. This has led to a proliferation of files each with fewer than 50
lines of code. Because of now needing to work with more files, I brought
in Browserify to manage including files and Gulp for managing the build
process.